added new ibslib, sleeping barrier, and RDF

+++ setup.py
@@ -1,6 +1,9 @@
-import os, shutil
+import os, shutil, platform
 from setuptools import find_packages
 from setuptools import setup
+python_version = float(platform.python_version()[:3])
+if python_version < 3.5:
+    raise Exception('Need python 3.5 or above')
 if os.path.exists('build'):
     shutil.rmtree('build')
 if os.path.exists('Genarris_2.0.egg-info'):
@@ -23,6 +26,11 @@ except ImportError:
 os.chdir(os.path.join('Genarris', 'ibslib'))
 os.system('python setup.py install')
 os.chdir(cwd)
+str_py_version = str(int(python_version * 10))
+torch_whl = 'https://download.pytorch.org/whl/cpu/torch-1.1.0-cp' + str_py_version + '-cp' + str_py_version + 'm-linux_x86_64.whl'
+os.system('pip install ' + torch_whl)
+torch_whl = 'https://download.pytorch.org/whl/cpu/torchvision-0.3.0-cp' + str_py_version + '-cp' + str_py_version + 'm-linux_x86_64.whl'
+os.system('pip install ' + torch_whl)
 if os.path.exists('ase'):
    shutil.rmtree('ase')
 os.system('pip uninstall -y ase')
